All Swift Data related code and views removed.
