--- Clef Notes/Views/AudioPlaybackCell.swift
@@ -1,114 +1,0 @@
-import SwiftUI
-import AVFoundation
-import SwiftData
-
-// A helper struct to make audio data transferable for the ShareLink.
-private struct AudioFile: Transferable {
-    let data: Data
-    let filename: String
-
-    static var transferRepresentation: some TransferRepresentation {
-        DataRepresentation(contentType: .mpeg4Audio) { audio in
-            audio.data
-        } importing: { data in
-            AudioFile(data: data, filename: "imported.m4a")
-        }
-        .suggestedFileName { audio in
-            audio.filename
-        }
-    }
-}
-
-// A unified view for playing ANY audio.
-struct AudioPlaybackCell: View {
-    let title: String
-    let subtitle: String
-    let data: Data?
-    let duration: TimeInterval?
-    let id: PersistentIdentifier
-    @ObservedObject var audioPlayerManager: AudioPlayerManager
-    
-    @State private var isScrubbing = false
-    
-    var isPlaying: Bool {
-      //  audioPlayerManager.currentlyPlayingID == id && audioPlayerManager.audioPlayer?.isPlaying == true
-        return false
-    }
-    
-    var body: some View {
-        VStack {
-            HStack {
-                VStack(alignment: .leading) {
-                    Text(title).font(.headline)
-                    Text(subtitle).font(.caption).foregroundColor(.secondary)
-                    if let duration = duration {
-                        Text("Duration: \(Int(duration))s").font(.caption2).foregroundColor(.gray)
-                    }
-                }
-                Spacer()
-                
-                // --- THIS IS THE FIX ---
-                // A new HStack provides better spacing and custom styling for the buttons.
-                HStack(spacing: 16) {
-                    if let audioData = data {
-                        ShareLink(
-                            item: AudioFile(data: audioData, filename: "\(title).m4a"),
-                            preview: SharePreview(title, image: Image(systemName: "waveform"))
-                        ) {
-                            Image(systemName: "square.and.arrow.up")
-                                .font(.title3)
-                                .frame(width: 40, height: 40)
-                                .background(Color(UIColor.systemGray6))
-                                .clipShape(Circle())
-                        }
-                    }
-                    
-                    Button(action: {
-                        if isPlaying {
-                            audioPlayerManager.stop()
-                        } else if data != nil {
-                            //audioPlayerManager.play(data: audioData, id: id)
-                        } else {
-                            print("Cannot play: audio data is nil")
-                        }
-                    }) {
-                        Image(systemName: isPlaying ? "stop.fill" : "play.fill")
-                            .font(.title3)
-                            .frame(width: 40, height: 40)
-                            .background(Color(UIColor.systemGray5))
-                            .clipShape(Circle())
-                    }
-                    .disabled(data == nil)
-                }
-                .buttonStyle(.plain) // Ensures the custom styling is applied correctly.
-            }
-            .padding(.vertical, 4)
-            
-            if isPlaying, let duration = duration {
-                Slider(
-                    value: Binding(
-                        get: { isScrubbing ? audioPlayerManager.currentTime : min(audioPlayerManager.currentTime, duration) },
-                        set: { isScrubbing = true; audioPlayerManager.currentTime = $0 }
-                    ),
-                    in: 0...duration,
-                    onEditingChanged: { editing in
-                        isScrubbing = editing
-                        if !editing {
-                            audioPlayerManager.seek(to: audioPlayerManager.currentTime)
-                        }
-                    }
-                )
-                .accentColor(.accentColor)
-                .padding(.horizontal, 8)
-                HStack {
-                    Text(String(format: "%02d:%02d", Int(audioPlayerManager.currentTime) / 60, Int(audioPlayerManager.currentTime) % 60))
-                        .font(.caption2).foregroundColor(.gray)
-                    Spacer()
-                    Text(String(format: "%02d:%02d", Int(duration) / 60, Int(duration) % 60))
-                        .font(.caption2).foregroundColor(.gray)
-                }
-                .padding(.horizontal, 8)
-            }
-        }
-    }
-}
